// web-client/src/graphql/types.ts
export type GetCampaignsQuery = {
  getCampaigns:  Array< {
    __typename: "Campaign",
    id: string,
    uid: string,
    created: number,
    updated: number,
    name: string,
    from: string,
    questionText: string,
    destinations:  Array< {
      __typename: "Destination",
      name: string,
      email: string,
    } >,
    stats:  {
      __typename: "CampaignStats",
      Created: number | null,
      Sent: number | null,
      Delivered: number | null,
      Complaint: number | null,
      Rejected: number | null,
      Bounce: number | null,
      Opened: number | null,
      Responded: number | null,
    },
  } >,
};
// ...
export type CampaignUpdateSubscription = {
  campaignUpdate:  {
    __typename: "Campaign",
    id: string,
    uid: string,
    created: number,
    updated: number,
    name: string,
    from: string,
    questionText: string,
    destinations:  Array< {
      __typename: "Destination",
      name: string,
      email: string,
    } >,
    stats:  {
      __typename: "CampaignStats",
      Created: number | null,
      Sent: number | null,
      Delivered: number | null,
      Complaint: number | null,
      Rejected: number | null,
      Bounce: number | null,
      Opened: number | null,
      Responded: number | null,
    },
  } | null,
};